move basic interactive functions into utils
